+++ activestorage/lib/active_storage/service/s3_service.rb
@@ -12,14 +12,15 @@ module ActiveStorage
     attr_reader :client, :bucket
     attr_reader :multipart_upload_threshold, :upload_options
 
-    def initialize(bucket:, upload: {}, public_service: false, **options)
+    def initialize(bucket:, upload: {}, **options)
+      @public = options[:public]
+      options.delete(:public)
+
       @client = Aws::S3::Resource.new(**options)
       @bucket = @client.bucket(bucket)
 
       @multipart_upload_threshold = upload.fetch(:multipart_threshold, 100.megabytes)
       @upload_options = upload
-
-      @public_service = public_service
     end
 
     def upload(key, io, checksum: nil, filename: nil, content_type: nil, disposition: nil, **)
@@ -93,7 +94,7 @@ module ActiveStorage
       { "Content-Type" => content_type, "Content-MD5" => checksum, "Content-Disposition" => content_disposition }
     end
 
-    protected
+    private
       def private_url(key, expires_in:, filename:, disposition:, content_type:)
         object_for(key).presigned_url :get, expires_in: expires_in.to_i,
           response_content_disposition: content_disposition_with(type: disposition, filename: filename),
@@ -104,7 +105,7 @@ module ActiveStorage
         object_for(key).public_url
       end
 
-    private
+
       MAXIMUM_UPLOAD_PARTS_COUNT = 10000
       MINIMUM_UPLOAD_PART_SIZE   = 5.megabytes
