+++ activestorage/test/controllers/disk_controller_test.rb
@@ -44,33 +44,6 @@ class ActiveStorage::DiskControllerTest < ActionDispatch::IntegrationTest
     assert_response :not_found
   end
 
-  test "showing public blob" do
-    with_service("public") do
-      blob = create_blob(content_type: "image/jpg")
-
-      get blob.url
-      assert_response :ok
-      assert_equal "image/jpg", response.headers["Content-Type"]
-      assert_equal "Hello world!", response.body
-    end
-  end
-
-  test "showing private blob as a public blob" do
-    blob = create_blob(content_type: "image/jpg")
-
-    with_service("public") do
-      get rails_disk_service_public_url(key: blob.key, filename: "foo.jpg")
-      assert_response :unauthorized
-    end
-  end
-
-  test "showing public blob with invalid key" do
-    with_service("public") do
-      get rails_disk_service_public_url(key: "abc123", filename: "foo.jpg")
-      assert_response :not_found
-    end
-  end
-
   test "directly uploading blob with integrity" do
     data = "Something else entirely!"
     blob = create_blob_before_direct_upload byte_size: data.size, checksum: Digest::MD5.base64digest(data)
