+++ activestorage/test/test_helper.rb
@@ -37,7 +37,7 @@ require "tmpdir"
 
 Rails.configuration.active_storage.service_configurations = SERVICE_CONFIGURATIONS.merge(
   "local" => { "service" => "Disk", "root" => Dir.mktmpdir("active_storage_tests") },
-  "public" => { "service" => "Disk", "root" => Dir.mktmpdir("active_storage_tests"), "public_service" => true },
+  "public" => { "service" => "Disk", "root" => Dir.mktmpdir("active_storage_tests"), "public" => true },
   "disk_mirror_1" => { "service" => "Disk", "root" => Dir.mktmpdir("active_storage_tests_1") },
   "disk_mirror_2" => { "service" => "Disk", "root" => Dir.mktmpdir("active_storage_tests_2") },
   "disk_mirror_3" => { "service" => "Disk", "root" => Dir.mktmpdir("active_storage_tests_3") },
